+++ tool/teleport-update/main.go
@@ -19,30 +19,23 @@
 package main
 
 import (
-	"bytes"
-	"compress/gzip"
 	"context"
-	"crypto/sha256"
 	"crypto/tls"
 	"crypto/x509"
-	"encoding/hex"
 	"errors"
-	"io"
 	"io/fs"
 	"log/slog"
 	"net/http"
 	"net/url"
 	"os"
+	"os/signal"
 	"path/filepath"
-	"runtime"
 	"syscall"
-	"text/template"
 	"time"
 
 	"github.com/google/renameio/v2"
 	"github.com/gravitational/trace"
 	"golang.org/x/net/http/httpproxy"
-	"golang.org/x/sys/unix"
 	"gopkg.in/yaml.v3"
 
 	"github.com/gravitational/teleport"
@@ -69,9 +62,13 @@ const (
 	templateEnvVar    = "TELEPORT_URL_TEMPLATE"
 	proxyServerEnvVar = "TELEPORT_PROXY"
 	updateGroupEnvVar = "TELEPORT_UPDATE_GROUP"
-)
 
-const cdnURITemplate = "https://cdn.teleport.dev/teleport-v{{.Version}}-{{.OS}}-{{.Arch}}-bin.tar.gz"
+	cdnURITemplate = "https://cdn.teleport.dev/teleport-v{{.Version}}-{{.OS}}-{{.Arch}}-bin.tar.gz"
+
+	versionsDirName = "versions"
+	updatesFileName = "updates.yaml"
+	lockFileName    = ".lock"
+)
 
 var plog = logutils.NewPackageLogger(teleport.ComponentKey, teleport.ComponentUpdater)
 
@@ -97,6 +94,7 @@ type cliConfig struct {
 func Run(args []string) error {
 	var ccfg cliConfig
 	ctx := context.Background()
+	ctx, _ = signal.NotifyContext(ctx, os.Interrupt, syscall.SIGTERM)
 
 	app := libutils.InitCLIParser("teleport-updater", appHelp).Interspersed(false)
 	app.Flag("debug", "Verbose logging to stdout.").Short('d').BoolVar(&ccfg.Debug)
@@ -129,7 +127,7 @@ func Run(args []string) error {
 		return trace.Errorf("setting up logger")
 	}
 
-	err = validate(&ccfg)
+	err = validateCLIConfig(&ccfg)
 	if err != nil {
 		return trace.Wrap(err)
 	}
@@ -138,9 +136,9 @@ func Run(args []string) error {
 	case enableCmd.FullCommand():
 		err = doEnable(ctx, &ccfg)
 	case disableCmd.FullCommand():
-		err = doDisable()
+		err = doDisable(ctx, &ccfg)
 	case updateCmd.FullCommand():
-		err = doUpdate(ctx)
+		err = doUpdate(ctx, &ccfg)
 	case versionCmd.FullCommand():
 		modules.GetModules().PrintVersion()
 	default:
@@ -168,27 +166,23 @@ func setupLogger(debug bool, format string) error {
 	return nil
 }
 
-// TODO: should be configurable?
-var (
-	versionsDir = filepath.Join(libdefaults.DataDir, "versions")
-	updatesYAML = filepath.Join(versionsDir, "updates.yaml")
-)
-
 const (
 	updatesVersion = "v1"
 	updatesKind    = "agent_versions"
 )
 
 type UpdatesConfig struct {
-	Version string `yaml:"version"`
-	Kind    string `yaml:"kind"`
-	Spec    struct {
-		Proxy         string `yaml:"proxy"`
-		Group         string `yaml:"group"`
-		URITemplate   string `yaml:"uri_template"`
-		Enabled       bool   `yaml:"enabled"`
-		ActiveVersion string `yaml:"active_version"`
-	} `yaml:"spec"`
+	Version string      `yaml:"version"`
+	Kind    string      `yaml:"kind"`
+	Spec    UpdatesSpec `yaml:"spec"`
+}
+
+type UpdatesSpec struct {
+	Proxy         string `yaml:"proxy"`
+	Group         string `yaml:"group"`
+	URITemplate   string `yaml:"uri_template"`
+	Enabled       bool   `yaml:"enabled"`
+	ActiveVersion string `yaml:"active_version"`
 }
 
 func readUpdatesConfig(path string) (*UpdatesConfig, error) {
@@ -233,7 +227,16 @@ func writeUpdatesConfig(filename string, cfg *UpdatesConfig) error {
 	return trace.Wrap(t.CloseAtomicallyReplace())
 }
 
-func doDisable() error {
+func doDisable(ctx context.Context, ccfg *cliConfig) error {
+	var (
+		versionsDir = filepath.Join(ccfg.DataDir, versionsDirName)
+		updatesYAML = filepath.Join(versionsDir, updatesFileName)
+	)
+	unlock, err := lock(filepath.Join(versionsDir, lockFileName))
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	defer unlock()
 	cfg, err := readUpdatesConfig(updatesYAML)
 	if err != nil {
 		return trace.Wrap(err)
@@ -249,7 +252,11 @@ func doDisable() error {
 }
 
 func doEnable(ctx context.Context, ccfg *cliConfig) error {
-	unlock, err := lock()
+	var (
+		versionsDir = filepath.Join(ccfg.DataDir, versionsDirName)
+		updatesYAML = filepath.Join(versionsDir, updatesFileName)
+	)
+	unlock, err := lock(filepath.Join(versionsDir, lockFileName))
 	if err != nil {
 		return trace.Wrap(err)
 	}
@@ -269,7 +276,13 @@ func doEnable(ctx context.Context, ccfg *cliConfig) error {
 		cfg.Spec.URITemplate = ccfg.Template
 	}
 	cfg.Spec.Enabled = true
-
+	if err := validateUpdatesSpec(&cfg.Spec); err != nil {
+		return trace.Wrap(err)
+	}
+	certPool, err := x509.SystemCertPool()
+	if err != nil {
+		return trace.Wrap(err)
+	}
 	addr, err := libutils.ParseAddr(cfg.Spec.Proxy)
 	if err != nil {
 		return trace.Errorf("failed to parse proxy server address: %w", err)
@@ -279,12 +292,29 @@ func doEnable(ctx context.Context, ccfg *cliConfig) error {
 		ProxyAddr: addr.Addr,
 		Timeout:   30 * time.Second,
 		Group:     cfg.Spec.Group,
+		Pool:      certPool,
 	})
 	if err != nil {
 		return trace.Errorf("failed to request version from proxy: %w", err)
 	}
+	resp.AgentVersion = "16.1.0"
 	if cfg.Spec.ActiveVersion != resp.AgentVersion {
-		err = createVersion(ctx, cfg.Spec.URITemplate, resp.AgentVersion)
+		client, err := newClient(&downloadConfig{
+			Pool: certPool,
+		})
+		if err != nil {
+			return trace.Wrap(err)
+		}
+		defer client.CloseIdleConnections()
+		tv := TeleportVersion{
+			VersionsDir:    filepath.Join(ccfg.DataDir, "versions"),
+			DownloadClient: client,
+		}
+		template := cfg.Spec.URITemplate
+		if template == "" {
+			template = cdnURITemplate
+		}
+		err = tv.Create(ctx, template, resp.AgentVersion)
 		if err != nil {
 			return trace.Wrap(err)
 		}
@@ -296,8 +326,12 @@ func doEnable(ctx context.Context, ccfg *cliConfig) error {
 	return nil
 }
 
-func doUpdate(ctx context.Context) error {
-	unlock, err := lock()
+func doUpdate(ctx context.Context, ccfg *cliConfig) error {
+	var (
+		versionsDir = filepath.Join(ccfg.DataDir, versionsDirName)
+		updatesYAML = filepath.Join(versionsDir, updatesFileName)
+	)
+	unlock, err := lock(filepath.Join(versionsDir, lockFileName))
 	if err != nil {
 		return trace.Wrap(err)
 	}
@@ -311,6 +345,10 @@ func doUpdate(ctx context.Context) error {
 		return trace.Errorf("updates disabled")
 	}
 
+	certPool, err := x509.SystemCertPool()
+	if err != nil {
+		return trace.Wrap(err)
+	}
 	addr, err := libutils.ParseAddr(cfg.Spec.Proxy)
 	if err != nil {
 		return trace.Errorf("failed to parse proxy server address: %w", err)
@@ -320,13 +358,30 @@ func doUpdate(ctx context.Context) error {
 		ProxyAddr: addr.Addr,
 		Timeout:   30 * time.Second,
 		Group:     cfg.Spec.Group,
+		Pool:      certPool,
 	})
 	if err != nil {
 		return trace.Errorf("failed to request version from proxy: %w", err)
 	}
 
+	resp.AgentVersion = "16.1.0"
 	if cfg.Spec.ActiveVersion != resp.AgentVersion {
-		err := createVersion(ctx, cfg.Spec.URITemplate, resp.AgentVersion)
+		client, err := newClient(&downloadConfig{
+			Pool: certPool,
+		})
+		if err != nil {
+			return trace.Wrap(err)
+		}
+		defer client.CloseIdleConnections()
+		tv := TeleportVersion{
+			VersionsDir:    filepath.Join(ccfg.DataDir, "versions"),
+			DownloadClient: client,
+		}
+		template := cfg.Spec.URITemplate
+		if template == "" {
+			template = cdnURITemplate
+		}
+		err = tv.Create(ctx, template, resp.AgentVersion)
 		if err != nil {
 			return trace.Wrap(err)
 		}
@@ -338,224 +393,10 @@ func doUpdate(ctx context.Context) error {
 	return nil
 }
 
-func download(ctx context.Context, url string) (r io.ReadSeekCloser, sum []byte, err error) {
-	f, err := os.CreateTemp("", "teleport-update-")
-	if err != nil {
-		return nil, nil, trace.Errorf("failed to create temporary file: %w", err)
-	}
-	defer func() {
-		if err != nil {
-			f.Close()
-			if err := os.Remove(f.Name()); err != nil {
-				plog.WarnContext(ctx, "Failed to cleanup temporary download after error", "error", err)
-			}
-		}
-	}()
-	free, err := freeDisk(os.TempDir())
-	if err != nil {
-		return nil, nil, trace.Errorf("failed to calculate free disk: %w", err)
-	}
-
-	client, err := newClient(&downloadConfig{})
-	if err != nil {
-		return nil, nil, trace.Wrap(err)
-	}
-	defer client.CloseIdleConnections()
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url, nil)
-	if err != nil {
-		return nil, nil, trace.Wrap(err)
-	}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, nil, trace.Wrap(err)
-	}
-	defer resp.Body.Close()
-
-	if resp.ContentLength < 0 {
-		plog.Warn("Content length missing from response, unable to verify Teleport download size")
-	} else if resp.ContentLength > free {
-		return nil, nil, trace.Errorf("size of download exceeds available disk space")
-	}
-	shaReader := sha256.New()
-	n, err := io.Copy(f, io.TeeReader(resp.Body, shaReader))
-	if err != nil {
-		return nil, nil, trace.Wrap(err)
-	}
-	if resp.ContentLength >= 0 && n != resp.ContentLength {
-		return nil, nil, trace.Errorf("mismatch in Teleport download size")
-	}
-	if _, err := f.Seek(0, io.SeekStart); err != nil {
-		return nil, nil, trace.Errorf("failed seek to start of download: %w", err)
-	}
-	return rmCloser{f}, shaReader.Sum(nil), nil
-}
-
-type rmCloser struct {
-	*os.File
-}
-
-func (r rmCloser) Close() error {
-	err := r.File.Close()
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	err = os.Remove(r.File.Name())
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	return nil
-}
-
-func getChecksum(ctx context.Context, url string) ([]byte, error) {
-	client, err := newClient(&downloadConfig{})
-	if err != nil {
+func lock(lockFile string) (func(), error) {
+	if err := os.MkdirAll(filepath.Dir(lockFile), 0755); err != nil {
 		return nil, trace.Wrap(err)
 	}
-	defer client.CloseIdleConnections()
-	ctx, cancel := context.WithTimeout(ctx, 30*time.Second)
-	defer cancel()
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url, nil)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-	defer resp.Body.Close()
-
-	r := io.LimitReader(resp.Body, 64)
-	var buf bytes.Buffer
-	_, err = io.Copy(&buf, r)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-	sum, err := hex.DecodeString(buf.String())
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-	return sum, nil
-}
-
-func removeVersion(version string) error {
-	versionPath := filepath.Join(versionsDir, version)
-	sumPath := filepath.Join(versionPath, "sha256")
-
-	// invalidate checksum first, to protect against partially-removed
-	// directory with valid checksum
-	if err := os.Remove(sumPath); err != nil {
-		return trace.Wrap(err)
-	}
-	if err := os.RemoveAll(versionPath); err != nil {
-		return trace.Wrap(err)
-	}
-	return nil
-}
-
-type TeleportVersion struct {
-	VersionsDir    string
-	DownloadClient *http.Client
-}
-
-func createVersion(ctx context.Context, uriTmpl, version string) error {
-	versionPath := filepath.Join(versionsDir, version)
-	sumPath := filepath.Join(versionPath, "sha256")
-
-	tmpl, err := template.New("uri").Parse(uriTmpl)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	var uriBuf bytes.Buffer
-	params := struct {
-		OS, Version, Arch string
-	}{runtime.GOOS, version, runtime.GOARCH}
-	err = tmpl.Execute(&uriBuf, params)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	uri := uriBuf.String()
-
-	sum, err := getChecksum(ctx, uri)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	existSum, err := readChecksum(sumPath)
-	if err == nil {
-		if bytes.Equal(existSum, sum) {
-			plog.InfoContext(ctx, "Version already present", "version", version)
-			return nil
-		}
-		plog.WarnContext(ctx, "Removing version that does not match checksum", "version", version)
-		if err := removeVersion(version); err != nil {
-			return trace.Wrap(err)
-		}
-	} else if !errors.Is(err, os.ErrNotExist) {
-		return trace.Wrap(err)
-	}
-	tgz, pathSum, err := download(ctx, uri)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	defer func() {
-		if err := tgz.Close(); err != nil {
-			plog.WarnContext(ctx, "Failed to cleanup temporary download after error", "error", err)
-		}
-	}()
-
-	if !bytes.Equal(sum, pathSum) {
-		return trace.Errorf("mismatched checksum, download possibly corrupt")
-	}
-	// avoid gzip bomb by validating checksum before decompression
-	n, err := uncompressedSize(tgz)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	if _, err := tgz.Seek(0, io.SeekStart); err != nil {
-		return trace.Errorf("failed seek to start: %w", err)
-	}
-	if err := os.MkdirAll(versionPath, 0755); err != nil {
-		return trace.Wrap(err)
-	}
-	free, err := freeDisk(versionPath)
-	if err != nil {
-		return trace.Errorf("failed to calculate free disk in %q: %w", versionPath, err)
-	}
-	if d := free - n; d < 0 {
-		return trace.Errorf("%q needs %d additional bytes of disk space for download", versionsDir, -d)
-	}
-	err = untar(tgz, versionPath)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	err = os.WriteFile(sumPath, sum, 0755)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	return nil
-}
-
-func readChecksum(path string) ([]byte, error) {
-	f, err := os.Open(path)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-	defer f.Close()
-	var buf bytes.Buffer
-	n, err := io.Copy(&buf, io.LimitReader(f, 65))
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-	if n != 64 {
-		return nil, trace.Errorf("mismatch in checksum size")
-	}
-	return buf.Bytes(), nil
-}
-
-func lock() (func(), error) {
-	// Build the path to the lock file that will be used by flock.
-	lockFile := filepath.Join(versionsDir, ".lock")
-
-	// Create the advisory lock using flock.
 	lf, err := os.OpenFile(lockFile, os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0666)
 	if err != nil {
 		return nil, trace.Wrap(err)
@@ -602,42 +443,16 @@ func newClient(cfg *downloadConfig) (*http.Client, error) {
 	}, nil
 }
 
-const reservedFreeDisk = 10_000_000 // 10 MiB
-
-func freeDisk(dir string) (int64, error) {
-	var stat unix.Statfs_t
-	err := unix.Statfs(dir, &stat)
-	if err != nil {
-		return 0, trace.Wrap(err)
+func validateCLIConfig(cfg *cliConfig) error {
+	if cfg.DataDir == "" {
+		cfg.DataDir = libdefaults.DataDir
 	}
-	if int64(stat.Bavail) < 0 {
-		return 0, trace.Errorf("invalid size")
-	}
-	return int64(stat.Bavail) - reservedFreeDisk, nil
+	return nil
 }
 
-func uncompressedSize(f io.Reader) (int64, error) {
-	// NOTE: The gzip length trailer is very unreliable,
-	//   but we could optimize this in the future if
-	//   we are willing to verify that all published
-	//   Teleport tarballs have valid trailers.
-	r, err := gzip.NewReader(f)
-	if err != nil {
-		return 0, trace.Wrap(err)
-	}
-	n, err := io.Copy(io.Discard, r)
-	if err != nil {
-		return 0, trace.Wrap(err)
-	}
-	return n, nil
-}
-
-func validate(cf *cliConfig) error {
-	if cf.Template == "" {
-		cf.Template = cdnURITemplate
-	}
-	if cf.DataDir == "" {
-		cf.DataDir = libdefaults.DataDir
+func validateUpdatesSpec(spec *UpdatesSpec) error {
+	if spec.Proxy == "" {
+		return trace.Errorf("proxy URL must be specified with --proxy or present in updates.yaml")
 	}
 	return nil
 }
